+++ vast-bindings/src/planet_type_type.rs
@@ -7,20 +7,14 @@ use spacetimedb_sdk::__codegen::{self as __sdk, __lib, __sats, __ws};
 #[derive(__lib::ser::Serialize, __lib::de::Deserialize, Clone, PartialEq, Debug)]
 #[sats(crate = __lib)]
 #[derive(Copy, Eq, Hash)]
-pub enum BuildingKind {
-    MiningDepot,
+pub enum PlanetType {
+    Solid,
 
-    Warehouse,
+    Ocean,
 
-    MilitaryGarrison,
-
-    SalesDepot,
-
-    ShipDepot,
-
-    Radar,
+    Gas,
 }
 
-impl __sdk::InModule for BuildingKind {
+impl __sdk::InModule for PlanetType {
     type Module = super::RemoteModule;
 }