+++ vast-bindings/src/scan_initiator_type.rs
@@ -7,20 +7,12 @@ use spacetimedb_sdk::__codegen::{self as __sdk, __lib, __sats, __ws};
 #[derive(__lib::ser::Serialize, __lib::de::Deserialize, Clone, PartialEq, Debug)]
 #[sats(crate = __lib)]
 #[derive(Copy, Eq, Hash)]
-pub enum BuildingKind {
-    MiningDepot,
-
-    Warehouse,
-
-    MilitaryGarrison,
-
-    SalesDepot,
-
-    ShipDepot,
-
+pub enum ScanInitiator {
     Radar,
+
+    Ship,
 }
 
-impl __sdk::InModule for BuildingKind {
+impl __sdk::InModule for ScanInitiator {
     type Module = super::RemoteModule;
 }